More boards

+++ rapport/rapport n°6/code_gr_24_part_6.py
@@ -79,7 +79,7 @@ def play_game(level_name, players_names, players_types, remote_id=None, remote_i
     total_turn = -1
 
     # The main game loop.
-    while game_running:        
+    while game_running:
         if total_turn > -1:
             write_log(game_data, u'It\'s turn nb %d' % (total_turn), log_info)
 
@@ -120,12 +120,13 @@ def play_game(level_name, players_names, players_types, remote_id=None, remote_i
         total_turn += 1
         game_running = is_game_continue(game_data)
 
-    show_game_board(game_data)
+        show_game_board(game_data)
 
     # Disconnect the remote player.
     if is_remote_game:
         disconnect_from_player(game_data['connection'])
 
+    raw_input()
     # Show the end game screen.
     show_end_game(game_data)
     raw_input()
@@ -307,8 +308,7 @@ def take_abandoned_ship(game_data):
                 if ship_owner == 'none':
                     abandoned_ships.append(ship)
                 else:
-                    if not ship_owner in owners:
-                        owners.append(ship_owner)
+                    owners.append(ship_owner)
 
         if len(owners) == 1:
             owner = owners[0]
@@ -580,9 +580,8 @@ def get_ai_input(game_data, player_name):
                 ship_order = get_battleship_action(game_data, ship_name, player_name)
         else:
             ship_order = get_battleship_action(game_data, ship_name, player_name)
-        
-        if ship_order != '':
-            ai_input += '%s:%s ' % (ship_name[len(player_name) + 1:], ship_order)
+
+        ai_input += '%s:%s ' % (ship_name[len(player_name) + 1:], ship_order)
 
     return ai_input[:-1]
 
@@ -797,7 +796,7 @@ def get_nearby_ship(game_data, target_ship, search_range):
     dy = -1
     nearby_ships = []
 
-    for i in range((search_range) ** 2):
+    for i in range((search_range * 2) ** 2):
         if (-search_range < x <= search_range) and (-search_range < y <= search_range) and abs(x) + abs(y) <= search_range:
 
             location = convert_coordinates((ship_location[0] + x, ship_location[1] + y), game_data['board_size'])
@@ -840,8 +839,6 @@ def follow_path(game_data, ship_name):
         to_do_node = ship['objective_path'].pop(0)
 
         return to_do_node['to_do']
-
-    return 'none'
 
 
 def get_closer(game_data, ship_name, objective):
@@ -921,6 +918,7 @@ def attack(game_data, ship):
     Implementation: Nicolas Van Bossuyt (v1. 19/03/17).
                     Bayron Mahy (v2. 22/03/17).
     """
+	#to comment
     ship_pos = game_data['ships'][ship]['location']
     ship_range = game_data['model_ship'][game_data['ships'][ship]['type']]['range']
     ship_owner = game_data['ships'][ship]['owner']
@@ -931,8 +929,7 @@ def attack(game_data, ship):
 
         for perhaps_target in nearby_ships:
             if game_data['ships'][perhaps_target]['owner'] != ship_owner and\
-               game_data['ships'][perhaps_target]['owner'] != 'none' and\
-               get_distance(ship_pos, predict_next_location(game_data, perhaps_target), game_data['board_size']) <= ship_range:
+               game_data['ships'][perhaps_target]['owner'] != 'none':
                 ships_targeted.append(perhaps_target)
 
         if len(ships_targeted) > 0:
@@ -1195,6 +1192,7 @@ def show_game_board(game_data):
 
     Implementation: Nicolas Van Bossuyt (v4. 10/02/17).
     """
+
     # Setup main canvas.
     screen_size = get_terminal_size()
     c = create_canvas(*screen_size)
@@ -1304,8 +1302,7 @@ def render_game_board(game_data):
 
             # Put the ship and the facing line on the game board.
             put_text(c, on_canvas_location[0] + 1, on_canvas_location[1], ship_icon, 1, 0, ship_color, None)
-            if ship_owner != 'none':
-                put_text(c, on_canvas_location[0] + 1 + ship_facing[0], on_canvas_location[1] + ship_facing[1],
+            put_text(c, on_canvas_location[0] + 1 + ship_facing[0], on_canvas_location[1] + ship_facing[1],
                      facing_char, 1, 0, ship_color, None)
 
         # There are more than one ship on the location, show how many there are.
@@ -1369,9 +1366,7 @@ def render_ship_list(game_data, width, height):
             put_text(c, 5, y, str(game_data['ships'][ship]['heal_points']))
 
             # Ship coordinates
-            ship_location = list(game_data['ships'][ship]['location'])
-            ship_location = (ship_location[1] + 1, ship_location[0] + 1)
-            put_text(c, 8, y, str(ship_location))
+            put_text(c, 9, y, str(game_data['ships'][ship]['location']))
 
             # Ship name.
             put_text(c, 17, y, ship[len(player + '_'):], 1, 0,
@@ -1521,7 +1516,8 @@ def parse_command(commands, player_name, game_data):
                         attack_location = ship_command.split('-')
                         if len(attack_location) == 2:
                             attack_location = (int(attack_location[1]) - 1, int(attack_location[0]) - 1)
-                            game_data['pending_attacks'].append((ship_name, game_data['ships'][ship_name]['location'], attack_location))
+                            game_data['pending_attacks'].append(
+                                (ship_name, game_data['ships'][ship_name]['location'], attack_location))
 
     return game_data
 
@@ -1664,29 +1660,25 @@ def command_attack(ship, ship_coordinate, target_coordinate, game_data):
     damages = ship_type['damages']
     distance = get_distance(ship_coordinate, target_coordinate, game_data['board_size'])
 
-    if distance <= ship_type['range']:
-        if len(game_data['board'][target_coordinate]) != 0:
-            game_data['nb_rounds'] = 0
+    if distance <= ship_type['range'] and len(game_data['board'][target_coordinate]) != 0:
+        game_data['nb_rounds'] = 0
 
-            # Give damages to all ships on targeted coordinate.
-            for target_ship in game_data['board'][target_coordinate]:
-                if  game_data['ships'][target_ship]['owner'] != 'none':
-                    # Give damages to the tageted ship.
-                    game_data['ships'][target_ship]['heal_points'] -= damages
+        # Give damages to all ships on targeted coordinate.
+        for target_ship in game_data['board'][target_coordinate]:
+            # Give damages to the tageted ship.
+            game_data['ships'][target_ship]['heal_points'] -= damages
 
-                    if game_data['ships'][target_ship]['heal_points'] <= 0:
-                        # Remove a space ship.
-                        write_log(game_data, '%s kill %s' % (ship, target_ship), log_warning)
-                        game_data['board'][target_coordinate].remove(target_ship)
-                        game_data['players'][game_data['ships'][target_ship]['owner']]['nb_ships'] -= 1
+            if game_data['ships'][target_ship]['heal_points'] <= 0:
+                write_log(game_data, '%s kill %s' % (ship, target_ship), log_warning)
 
-                        del game_data['ships'][target_ship]
-                    else:
-                        write_log(game_data, '%s shot %s' % (ship, target_ship), log_warning)
-        else:
-            write_log(game_data, '%s shoot failed %s!' % (ship, str(target_coordinate)), log_warning)
-    else:
-        write_log(game_data, '%s shoot out of range !' % ship, log_error)
+                # Remove a space ship.
+                game_data['board'][target_coordinate].remove(target_ship)
+                if game_data['ships'][target_ship]['owner'] != 'none':
+                    game_data['players'][game_data['ships'][target_ship]['owner']]['nb_ships'] -= 1
+
+                del game_data['ships'][target_ship]
+            else:
+                write_log(game_data, '%s shot %s' % (ship, target_ship), log_warning)
 
     return game_data
 
@@ -1711,18 +1703,13 @@ def get_distance(coord1, coord2, size):
                    Bayron Mahy (v2. 19/03/17)
     Implementation: Nicolas Van Bossuyt, Alisson Leist (v1. 14/02/17)
                     Nicolas Van Bossuyt (v2. 09/03/17)
-                    Nicolas Van Bossuyt (v3. 03/05/17)
     """
 
     def distance(a, b, size):
-        # size -= 1
-    
-        if a > b:
-            a, b = b, a
 
+        size -= 1
         if abs(a - b) > size / 2:
             a += size
-
         return abs(a - b)
 
     return distance(coord1[0], coord2[0], size[0]) + distance(coord1[1], coord2[1], size[1])
@@ -2116,16 +2103,13 @@ def write_log(game_data, message, type=0):
     Specification: Nicolas Van Bossuyt (v1. 18/04/2017)
     Implementation: Nicolas Van Bossuyt (v1. 18/04/2017)
     """
-    prefix = ['INFO', 'WARN', 'ERRO', 'INPU']
+
     game_data['game_logs'].append((type, message))
-
-
-
     f = open('log.txt', 'r')
     txt = f.read()
     f.close()
 
-    txt += '%s - %s\n' % (prefix[type], message)
+    txt += '%s\n' % message
 
     f = open('log.txt', 'w')
     f.write(txt)
@@ -2162,4 +2146,4 @@ def dict_sort(items, key):
 # Use for quick debuging.
 
 if __name__ == '__main__':
-    play_game('board/test.cis', ('NicolasLeRebelDeL\'espace', 'A.I.C.I.S.', 'bot', 'botbot'), ('ai','ai','ai','ai'))
+    play_game('board/crops_circle.cis', ('NicolasLeRebelDeL\'espace', 'A.I.C.I.S.', 'bot', 'botbot'), ('ai','ai','ai','ai'))
